requirements.txt
# Your requirements go here
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gspread==3.7.0